requirements.txt

flask==3.0.0
pandas==2.2.2
openpyxl==3.1.5
playwright==1.46.0
python-dotenv==1.0.1
